# Speed up two slow operations, which were unnecessarily searching all …