…roads linearly.

Correctness: Tests stay the same, but we're not currently checking anything that turn restrictions would affect.

I had an enormous map with ~300,000 roads that took 93s to collapse short roads! I don't have access to the computer processing that right now, but I found a smaller map that took 15s each for the two operations (collapsing short roads and degenerate intersections). With this PR, drops down to 0.3s and 2s, so I'll consider this a good-enough win.